requirements.txt
rottentomatoes-python>=0.5.18
letterboxdpy>=3.15

#csm:
beautifulsoup4>=4.11.1

#imdb:
requests>=2.28.1

#imdbpy:
#atomicwrites>=1.1.5
#attrs>=18.1.0
#lxml>=4.2.3
#more-itertools>=4.2.0
#packaging>=17.1
#pluggy>=0.7.1
#py>=1.5.4
#pyparsing>=2.2.0
#pytest>=3.6.4
#six>=1.11.0
#SQLAlchemy>=1.3.0
#tox>=3.1.2
#virtualenv>=16.0.0
